Feature/fix

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,18 +1,18 @@
 package ImageHoster.controller;
 
+import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
+import ImageHoster.service.CommentService;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -26,6 +26,9 @@ public class ImageController {
 
     @Autowired
     private TagService tagService;
+
+    @Autowired
+    private CommentService commentService;
 
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
@@ -45,11 +48,19 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+    @RequestMapping("/images/{id}/{title}")
+    public String showImage(@PathVariable("title") String title, Model model, @PathVariable("id") int id,HttpSession session) {
+        Image image = imageService.getImage(id);
+                //imageService.getImageByTitle(title,id);
+        User user = (User) session.getAttribute("loggeduser");
+        Comment comment=new Comment();
+//        comment.setUser(user);
+//        comment.setImage(image);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+
+        model.addAttribute("Comment",comment);
+        model.addAttribute("comments",commentService.getAllComments(id));
         return "images/image";
     }
 
@@ -92,12 +103,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
         Image image = imageService.getImage(imageId);
 
+        User user = (User) session.getAttribute("loggeduser");
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
+        if(user.getId()!=image.getUser().getId()) {
+           // redirectAttributes.addFlashAttribute("editError", "Only the owner of the image can edit the image");
+            model.addAttribute("editError", "Only the owner of the image can edit the image");
+            return "forward:/images/" + imageId+"/"+image.getTitle();
+            //return "redirect:/images/" + imageId+"/"+image.getTitle();
+
+        }
+
         return "images/edit";
     }
 
@@ -127,12 +147,14 @@ public class ImageController {
 
         updatedImage.setId(imageId);
         User user = (User) session.getAttribute("loggeduser");
+
         updatedImage.setUser(user);
         updatedImage.setTags(imageTags);
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+
+        return "redirect:/images/" + imageId+"/"+updatedImage.getTitle();
     }
 
 
@@ -140,7 +162,18 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session, RedirectAttributes redirectAttributes,Model model) {
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        if(user.getId()!=image.getUser().getId()) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            model.addAttribute("deleteError", "Only the owner of the image can delete the image");
+           // redirectAttributes.addFlashAttribute("deleteError", "Only the owner of the image can delete the image");
+            return "forward:/images/" + imageId+"/"+image.getTitle();
+
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
@@ -182,9 +215,13 @@ public class ImageController {
             tagString.append(tags.get(i).getName()).append(",");
         }
 
-        Tag lastTag = tags.get(tags.size() - 1);
-        tagString.append(lastTag.getName());
+        try {
+            Tag lastTag = tags.get(tags.size() - 1);
+            tagString.append(lastTag.getName());
 
-        return tagString.toString();
+            return tagString.toString();
+        }catch (Exception e){
+            return tagString.toString();
+        }
     }
 }
